# src/api.yaml
openapi: 3.0.0
info:
  title: DreamTrip API
  description: API for authentication user and administration
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
paths:
  /api/v1/user/save:
    post:
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                type: string
                example: "User created successfully"
  /api/v1/user/login:
    post:
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDTO'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginMessage'
  /api/v1/admin/login:
    post:
      summary: Authenticate administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginDTO'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginMessage'
  /api/v1/user/info:
    get:
      summary: Get user information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
  /api/v1/user/login-history:
    get:
      summary: Get user login history
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: username
          schema:
            type: string
          required: true
          description: The username to retrieve login history for
      responses:
        '200':
          description: User login history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserLoginHistoryDTO'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: string
                example: "User with username {username} not found"
  /api/v1/user/logout:
    post:
      summary: Logout user
      responses:
        '200':
          description: User logged out successfully
          content:
            application/json:
              schema:
                type: string
                example: "User logged out successfully"
components:
  schemas:
    UserDTO:
      type: object
      properties:
        username:
          type: string
        userLastName:
          type: string
        email:
          type: string
        password:
          type: string
        phoneNumber:
          type: string
      required:
        - username
        - userLastName
        - email
        - password
    LoginDTO:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AdminLoginDTO:
      type: object
      properties:
        email:
          type: string
        key:
          type: string
        password:
          type: string
      required:
        - email
        - key
        - password
    LoginMessage:
      type: object
      properties:
        message:
          type: string
        status:
          type: boolean
      required:
        - message
        - status
    AdminLoginMessage:
      type: object
      properties:
        message:
          type: string
        status:
          type: boolean
      required:
        - message
        - status
    UserInfo:
      type: object
      properties:
        username:
          type: string
        userLastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
      required:
        - username
        - userLastName
        - email
        - phoneNumber
        - address
    UserLoginHistoryDTO:
      type: object
      properties:
        id:
          type: integer
        loginTime:
          type: string
          format: date-time
      required:
        - id
        - loginTime