# src/main/java/com/hei/dreamtip/dto/UserDTO.java
package com.hei.dreamtip.dto;

public class UserDTO {
    private int userId;
    private String username;
    private String userLastName;
    private String email;
    private String password;
    private String phoneNumber;
    private String address; // Added address field

    public UserDTO() {
        this.userId = 0; // You can set default values for integer fields, assuming 0 is appropriate
        this.username = "";
        this.userLastName = "";
        this.email = "";
        this.password = "";
        this.phoneNumber = "";
        this.address = "";
    }

    public UserDTO(int userId, String username, String userLastName, String email, String password, String phoneNumber, String address) {
        this.userId = userId;
        this.username = username;
        this.userLastName = userLastName;
        this.email = email;
        this.password = password;
        this.phoneNumber = phoneNumber;
        this.address = address; // Initialize address
    }

    public int getUserId() {
        return userId;
    }

    public void setUserId(int userId) {
        this.userId = userId;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getUserLastName() {
        return userLastName;
    }

    public void setUserLastName(String userLastName) {
        this.userLastName = userLastName;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getPhoneNumber() {
        return phoneNumber;
    }

    public void setPhoneNumber(String phoneNumber) {
        this.phoneNumber = phoneNumber;
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    @Override
    public String toString() {
        return "UserDTO{" +
                "userId=" + userId +
                ", username='" + username + '\'' +
                ", userLastName='" + userLastName + '\'' +
                ", email='" + email + '\'' +
                ", password='" + password + '\'' +
                ", phoneNumber='" + phoneNumber + '\'' +
                ", address='" + address + '\'' +
                '}';
    }

    // Getters and setters...
}